+++ Firebase/InstanceID/FIRInstanceID.m
@@ -385,8 +385,11 @@ static FIRInstanceID *gInstanceID;
 
   FIRInstanceIDDeleteTokenHandler newHandler = ^(NSError *error) {
     // If a default token is deleted successfully, reset the defaultFCMToken too.
-    if (!error && [self isDefaultTokenWithAuthorizedEntity:authorizedEntity scope:scope]) {
-      self.defaultFCMToken = nil;
+    if (!error) {
+      if ([self isDefaultTokenWithAuthorizedEntity:authorizedEntity scope:scope] ||
+          [authorizedEntity isEqualToString:@"*"]) {
+        self.defaultFCMToken = nil;
+      }
     }
     dispatch_async(dispatch_get_main_queue(), ^{
       handler(error);
@@ -523,7 +526,7 @@ static FIRInstanceID *gInstanceID;
         return;
       }
 
-      [self.tokenManager.authService resetCheckinWithHandler:^(NSError *error) {
+      [self deleteCheckinWithHandler:^(NSError *_Nullable error) {
         if (error) {
           if (handler) {
             handler(error);
@@ -549,6 +552,10 @@ static FIRInstanceID *gInstanceID;
 }
 
 #pragma mark - Checkin
+
+- (void)deleteCheckinWithHandler:(void (^)(NSError *error))handler {
+  [self.tokenManager.authService resetCheckinWithHandler:handler];
+}
 
 - (BOOL)tryToLoadValidCheckinInfo {
   FIRInstanceIDCheckinPreferences *checkinPreferences =