+++ scripts/release_testing_setup.sh
@@ -12,7 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-set -x
+set -xe
+
+TESTINGMODE=${1-}
 
 if [ -f "${HOME}/.cocoapods/repos" ]; then
   find "${HOME}/.cocoapods/repos" -type d -maxdepth 1 -exec sh -c 'pod repo remove $(basename {})' \;
@@ -20,12 +22,33 @@ fi
 git config --global user.email "google-oss-bot@example.com"
 git config --global user.name "google-oss-bot"
 mkdir -p /tmp/test/firebase-ios-sdk
-git clone -b "${podspec_repo_branch}" https://github.com/firebase/firebase-ios-sdk.git "${local_sdk_repo_dir}"
-cd  "${local_sdk_repo_dir}"
-git tag -a "test" -m "release testing"
-# Update source and tag, e.g.  ":git => 'https://github.com/firebase/firebase-ios-sdk.git'" to
-# ":git => /tmp/test/firebase-ios-sdk"
-sed  -i "" "s/\s*:git.*/:git => '${local_sdk_repo_dir}',/; s/\s*:tag.*/:tag => 'test'/" *.podspec
+echo "git clone ${podspec_repo_branch} from github.com/firebase/firebase-ios-sdk.git to ${local_sdk_repo_dir}"
+set +x
+git clone -q -b "${podspec_repo_branch}" https://"${BOT_TOKEN}"@github.com/firebase/firebase-ios-sdk.git "${local_sdk_repo_dir}"
+set -x
+
+if [ "$TESTINGMODE" = "nightly_testing" ]; then
+  tag_version="nightly-test-${test_version}"
+  echo "A new tag, ${tag_version},for nightly release testing will be created."
+fi
+if [ "$TESTINGMODE" = "RC_testing" ]; then
+  tag_version="CocoaPods-${test_version}.nightly"
+  echo "A new tag, ${tag_version},for prerelease testing will be created."
+fi
+# Update a tag.
+if [ -n "$tag_version" ]; then
+  cd  "${local_sdk_repo_dir}"
+  set +e
+  # If tag_version is new to the remote, remote cannot delete a non-existent
+  # tag, so error is allowed here.
+  git push origin --delete "${tag_version}"
+  set -e
+  git tag -f -a "${tag_version}" -m "release testing"
+  git push origin "${tag_version}"
+  # Update source and tag, e.g.  ":tag => 'CocoaPods-' + s.version.to_s" to
+  # ":tag => test"
+  sed  -i "" "s/\s*:tag.*/:tag => '${tag_version}'/" *.podspec
+fi
 cd "${GITHUB_WORKSPACE}/ZipBuilder"
 swift build
 # Update Pod versions.