+++ GoogleDataTransport/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- Legacy pre Xcode 10 compatibility checks removed. (#6486)
+
 # v7.4.0
 - Limit disk space consumed by GoogleDataTransport to store events. (#6365)
 - Fix `GDTTransformer` background task handling.  (#6258)